Support for Analytics/Map/Dashboard integration
