+++ src/app/analytics/time/timeseries-analytics-form.component.scss
