+++ src/app/analytics/time/timeseries-analytics.component.scss
