--- src/app/analytics/widgets/twitter-timeseries/timeseries-config-form.component.html
@@ -1,49 +1,0 @@
-<form class="ttc-form" (keydown.enter)="$event.preventDefault()">
-    <mat-button-toggle-group *ngIf="this.pref.featureSupported('dashboard')" name="eoc" class="eoc"
-                             aria-label="Exceedance or Count" [(ngModel)]="data.state.eoc">
-        <mat-button-toggle value="exceedance">Exceedance</mat-button-toggle>
-        <mat-button-toggle value="count">Count</mat-button-toggle>
-    </mat-button-toggle-group>
-    <mat-form-field *ngIf="data.component.textFilter" class="text-search" >
-        <mat-label>Filter by Text</mat-label>
-        <mat-icon matPrefix>search</mat-icon>
-        <input #textSearchInput [formControl]="searchControl" matInput placeholder="Text search, e.g. coastal">
-        <button mat-button mat-icon-button matSuffix (click)="clearTextSearch()" *ngIf="data.state.textSearch">
-            <mat-icon>close</mat-icon>
-        </button>
-
-    </mat-form-field>
-    <mat-form-field *ngIf="data.component.regionFilter" class="region-chooser" >
-        <mat-label>Filter by Region</mat-label>
-        <mat-chip-list #chipList aria-label="Topic selection">
-            <mat-chip
-                    [class]="'region-level-'+region.level"
-                    *ngFor="let region of regions"
-                    [selectable]="true"
-                    [removable]="data.component.removable"
-                    [value]="region"
-                    (removed)="remove(region)">
-                {{region.text}}
-                <mat-icon matChipRemove *ngIf="data.component.removable">cancel</mat-icon>
-            </mat-chip>
-            <input
-                    placeholder="Add regions ..."
-                    #regionInput
-                    [formControl]="regionControl"
-                    [matAutocomplete]="auto"
-                    [matChipInputFor]="chipList"
-                    [matChipInputSeparatorKeyCodes]="separatorKeysCodes"
-                    (matChipInputTokenEnd)="add($event)">
-        </mat-chip-list>
-        <mat-autocomplete #auto="matAutocomplete" (optionSelected)="selected($event)">
-            <mat-option *ngFor="let region of filteredRegions | async" [value]="region">
-                {{region.text}}
-            </mat-option>
-        </mat-autocomplete>
-        <button mat-button mat-icon-button matSuffix (click)="clearRegions()" *ngIf="filteredRegions">
-            <mat-icon>close</mat-icon>
-        </button>
-
-    </mat-form-field>
-
-</form>
