--- src/app/analytics/widgets/twitter-timeseries/timeseries-config-form.component.scss
@@ -1,54 +1,0 @@
-.text-search {
-    min-width: 200px;
-    max-width: 400px;
-    margin-right: 48px;
-
-}
-
-.region-chooser {
-    min-width: 300px;
-    max-width: 600px;
-}
-
-.eoc {
-    margin-right: 48px;
-    min-width: 160px;
-    max-width: 176px;
-}
-
-form.ttc-form > * {
-    margin-bottom: 0px !important;
-}
-
-.region-level-0 {
-    background-color: hsl(216, 30%, 60%);
-    color: white !important;
-}
-.region-level-1 {
-    background-color: hsl(216, 50%, 50%);
-    color:white !important;
-}
-.region-level-2 {
-    background-color: hsl(216, 70%, 30%);
-    color: white !important;
-}
-
-::ng-deep {
-
-
-    div.mat-chip-list-wrapper {
-        max-height: 78px;
-        overflow: auto;
-
-        .mat-chip {
-            flex: 0 0 auto;
-
-
-        }
-    }
-
-    .mat-chip-list {
-        margin-left: 20px;
-    }
-
-}
