--- src/app/analytics/widgets/twitter-timeseries/twitter-timeseries.component.scss
@@ -1,14 +1,0 @@
-div.timeseries-chart-spinner-container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 100%;
-    position: absolute;
-}
-
-.dct-containing {
-    margin-right:2em;
-}
-.dct-containing-term {
-    font-weight: bolder;
-}
