+++ src/app/app.module.ts
@@ -37,18 +37,17 @@ import {DashboardComponent} from "./analytics/dashboard/dashboard.component";
 import {TimeSeriesChartComponent} from "./charts/timeseries-chart/timeseries-chart.component";
 import {BarChartComponent} from "./charts/bar-chart/bar-chart.component";
 import {PieChartComponent} from "./charts/pie-chart/pie-chart.component";
-import {TwitterTimeseriesComponent} from "./analytics/widgets/twitter-timeseries/twitter-timeseries.component";
 import {AmplifyUIAngularModule} from "@aws-amplify/ui-angular";
 import {TweetCopyDialogComponent} from "./map/twitter/tweet-list/tweet-copy-dialog/tweet-copy-dialog.component";
 import {StripHtmlPipe} from "./strip.pipe";
-import {TimeseriesConfigFormComponent} from "./analytics/widgets/twitter-timeseries/timeseries-config-form.component";
 import {TimeSeriesMultiChartComponent} from "./charts/timeseries-multi-chart/timeseries-multi-chart.component";
 import { MapGraphSidebarComponent } from "./map/map-graph-sidebar/map-graph-sidebar.component";
 import { AnalyticsComponent } from "./analytics/analytics.component";
-import {TimeseriesAnalyticsComponent} from "./analytics/twitter-timeseries-analytics/timeseries-analytics.component";
-import {TimeseriesAnalyticsFormComponent} from "./analytics/twitter-timeseries-analytics/timeseries-analytics-form.component";
-import {TimeSeriesMultipleQueryChartComponent} from "./analytics/twitter-timeseries-analytics/timeseries-multi-chart/time-series-multiple-query-chart.component";
-import { SaveGraphDialogComponent } from './analytics/twitter-timeseries-analytics/save-graph-dialog/save-graph-dialog.component';
+import {TimeseriesAnalyticsComponent} from "./analytics/time/timeseries-analytics.component";
+import {TimeseriesAnalyticsFormComponent} from "./analytics/time/timeseries-analytics-form.component";
+import {TimeSeriesMultipleQueryChartComponent} from "./analytics/time/timeseries-multi-chart/time-series-multiple-query-chart.component";
+import { SaveGraphDialogComponent } from "./analytics/time/save-graph-dialog/save-graph-dialog.component";
+import {TimeseriesWidgetComponent} from "./analytics/widgets/twitter-timeseries/timeseries-widget.component";
 
 @NgModule({
               declarations:    [
@@ -77,14 +76,14 @@ import { SaveGraphDialogComponent } from './analytics/twitter-timeseries-analyti
                   TimeSeriesChartComponent,
                   BarChartComponent,
                   PieChartComponent,
-                  TwitterTimeseriesComponent, TimeseriesConfigFormComponent,
                   TimeSeriesMultiChartComponent,
                   MapGraphSidebarComponent,
                   AnalyticsComponent,
                   TimeseriesAnalyticsComponent,
                   TimeseriesAnalyticsFormComponent,
                   TimeSeriesMultipleQueryChartComponent,
-                  SaveGraphDialogComponent
+                  SaveGraphDialogComponent,
+                  TimeseriesWidgetComponent
               ],
               imports:         [
                   BrowserModule,
