+++ src/app/map/map-graph-sidebar/map-graph-sidebar.component.scss
@@ -1,7 +1,6 @@
 .graph-wrapper {
     margin: 12px;
     margin-top:24px;
-    max-height: 380px;
     max-width:460px;
 }
