+++ src/app/map/map-graph-sidebar/map-graph-sidebar.component.ts
@@ -4,56 +4,50 @@ import {MatDialog} from "@angular/material/dialog";
 import {DashboardService} from "../../pref/dashboard.service";
 import {Logger} from "@aws-amplify/core";
 import {PreferenceService} from "../../pref/preference.service";
+import {EOC, TimeseriesAnalyticsComponentState} from "../../analytics/timeseries";
+import {Router} from "@angular/router";
+
 const log = new Logger("map-graph-sidebar");
 
 @Component({
-               selector:    "app-map-graph-sidebar",
-               templateUrl: "./map-graph-sidebar.component.html",
-               styleUrls:   ["./map-graph-sidebar.component.scss"]
+             selector:    "app-map-graph-sidebar",
+             templateUrl: "./map-graph-sidebar.component.html",
+             styleUrls:   ["./map-graph-sidebar.component.scss"]
            })
 export class MapGraphSidebarComponent implements OnInit {
-    @Input() selection: RegionSelection;
-    @ViewChild("expandedCountGraph") expandedCountGraph: TemplateRef<any>;
-    @ViewChild("expandedExceedanceGraph") expandedExceedanceGraph: TemplateRef<any>;
+  @Input() selection: RegionSelection;
 
-    public regionList: string[] = [];
+  public regionList: string[] = [];
 
-    constructor(public dialog: MatDialog, public dash: DashboardService, public pref:PreferenceService) { }
+  constructor(public dialog: MatDialog, public dash: DashboardService, public pref: PreferenceService,
+              protected _router: Router) { }
 
-    ngOnInit(): void {
-        this.regionList = this.selection.regionNames();
-        this.selection.changed.subscribe(i => this.regionList = this.selection.regionNames());
+  ngOnInit(): void {
+    this.regionList = this.selection.regionNames();
+    this.selection.changed.subscribe(i => this.regionList = this.selection.regionNames());
+  }
+
+  public expandCountGraph() {
+    this._router.navigate(["/analytics/time"], {queryParams: {region: this.regionList, eoc: "count"}});
+  }
+
+  public expandExceedanceGraph() {
+    this._router.navigate(["/analytics/time"], {queryParams: {region: this.regionList, eoc: "exceedance"}});
+  }
+
+  public showDashboard() {
+    this._router.navigate(["/dashboard"]);
+  }
+
+  public graphStateForRegions(eoc: EOC): TimeseriesAnalyticsComponentState {
+    const state: TimeseriesAnalyticsComponentState = {queries: [], eoc, lob: "line"};
+    for (const region of this.regionList) {
+      state.queries.push({textSearch: "", regions: [region]});
     }
+    return state;
+  }
 
-    public expandCountGraph() {
-        const dialogRef = this.dialog.open(this.expandedCountGraph, {
-            width:  "80vw",
-            height: "80vh",
-            minWidth: "800px",
-            minHeight: "800px",
-            maxHeight: "1024px"
-        });
-
-        dialogRef.afterClosed().subscribe(() => {
-            log.debug("The dialog was closed");
-        });
-    }
-
-    public expandExceedanceGraph() {
-        const dialogRef = this.dialog.open(this.expandedExceedanceGraph, {
-            width:  "80vw",
-            height: "80vh",
-            minWidth: "800px",
-            minHeight: "800px",
-            maxHeight: "1024px"
-        });
-
-        dialogRef.afterClosed().subscribe(() => {
-            log.debug("The dialog was closed");
-        });
-    }
-
-    public showDashboard() {
-        window.open("/dashboard", "___dashboard___");
-    }
+  public regionText() {
+    return this.selection.asTitleText();
+  }
 }
